Block-allocation VRAM pool type is now functional. It's the default type used for buffers.

+++ Electricity2/MemUtils.h
@@ -3,8 +3,8 @@
 
 namespace Electricity::Utils::Memory
 {
-	constexpr uint64 ToKilobytes( const uint32 uNumKb ) noexcept { return uNumKb * 1024; }
-	constexpr uint64 ToMegabytes( const uint32 uNumMb ) noexcept { return uNumMb * 1024 * 1024; }
+	constexpr uint64 ToKilobytes( const uint64 uNumKb ) noexcept { return uNumKb * 1024; }
+	constexpr uint64 ToMegabytes( const uint64 uNumMb ) noexcept { return uNumMb * 1024 * 1024; }
 
 	void Clear( void* pMem, const uint32 uSize ) noexcept;
 }
